% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fdrcorr}
\alias{fdrcorr}
\title{FDR correction for multiple testing}
\usage{
fdrcorr(test_list)
}
\arguments{
\item{test_list}{A numeric vector of p-values to be corrected.}
}
\value{
A numeric vector of FDR-adjusted p-values (q-values).
}
\description{
Performs False Discovery Rate (FDR) correction using the Benjamini-Hochberg method.
This is a wrapper around \code{\link{p.adjust}} with method = "fdr".
}
