% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{get_cellLine_list}
\alias{get_cellLine_list}
\title{Determine training and test cell-lines for a given drug}
\usage{
get_cellLine_list(
  infunc_cancerType = "PanCan",
  infunc_drugName,
  exclude_cancer = "PanCan",
  infunc_response,
  force_add_cellLines = TRUE,
  force_add_cellLines_list = NA
)
}
\arguments{
\item{infunc_cancerType}{Character. Cancer type for training. Default = "PanCan".}

\item{infunc_drugName}{Character. Name of the drug.}

\item{exclude_cancer}{Character. Cancer type to exclude from training. Default = "PanCan".}

\item{infunc_response}{Named numeric vector. Drug response data.}

\item{force_add_cellLines}{Logical. Whether to force add additional cell lines. Default = TRUE.}

\item{force_add_cellLines_list}{Character vector. Cell line IDs to force add. Default = NA.}
}
\value{
A list of length 2:
\itemize{
\item Element 1: common_cellLines - Cell lines to use for training
\item Element 2: cellLines2remove - Cell lines excluded (reserved for testing)
}
}
\description{
Identifies which cell lines should be used for training and which should be
excluded (reserved for testing) based on cancer type and single-cell data availability.
}
