% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_perception_mirrors}
\alias{get_perception_mirrors}
\title{Get current download mirrors for PERCEPTION}
\usage{
get_perception_mirrors()
}
\value{
Character vector of mirror URLs.
}
\description{
Get current download mirrors for PERCEPTION
}
\examples{
get_perception_mirrors()
}
