% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{package}
\name{plot_perception}
\alias{PERCEPTION}
\alias{PERCEPTION-package}
\alias{plot_perception}
\title{PERCEPTION Visualization Functions}
\description{
This module provides visualization functions for PERCEPTION model results,
including t-SNE/UMAP plots, clone distribution plots, ROC curves, and more.
These functions are designed to work seamlessly with trained PERCEPTION models
and patient/single-cell expression data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SunPast/PERCEPTION}
  \item Report bugs at \url{https://github.com/SunPast/PERCEPTION/issues}
}

}
\author{
\strong{Maintainer}: Jia Ding \email{jiading682@gmail.com}

Authors:
\itemize{
  \item Jia Ding \email{jiading682@gmail.com}
}

}
