% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{range01}
\alias{range01}
\title{Change range to 0-1}
\usage{
range01(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector scaled between 0 and 1.
}
\description{
Scales a numeric vector to the 0-1 range using the 5th and 95th percentiles
as thresholds to handle outliers.
}
