% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rank_normalization_mat}
\alias{rank_normalization_mat}
\title{Rank-normalize each column of a metrix}
\usage{
rank_normalization_mat(mat)
}
\arguments{
\item{mat}{A numeric matrix. Rows = genes, columns = cells.}
}
\value{
A rank-normalized matrix.
}
\description{
Convert each column to ranks and divide by column length.
}
