% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{run_parallel_feature_ranking_bulk}
\alias{run_parallel_feature_ranking_bulk}
\title{Parallel feature ranking for multiple drugs}
\usage{
run_parallel_feature_ranking_bulk(
  infunc_DrugsToUse,
  id_cancerType = "PanCan",
  infunc_exclude_cancer = "PanCan",
  infunc_GOI,
  ncores = 4
)
}
\arguments{
\item{infunc_DrugsToUse}{Character vector. Drug names to rank features for.}

\item{id_cancerType}{Character. Cancer type. Default = "PanCan".}

\item{infunc_exclude_cancer}{Character. Cancer type to exclude. Default = "PanCan".}

\item{infunc_GOI}{Character vector. Genes of Interest.}

\item{ncores}{Integer. Number of cores for parallel processing. Default = 4.}
}
\value{
A list of feature ranking results, one per drug.
}
\description{
Runs feature_ranking_bulk in parallel for a list of drugs.
}
