% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stripall2match}
\alias{stripall2match}
\title{Strip non-alphanumeric characters and lowercase}
\usage{
stripall2match(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Lowercase character vector with non-alphanumeric characters removed.
}
\description{
Primarily used to facilitate inconsistent naming comparison (e.g., drug names).
}
