% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strsplit_customv0}
\alias{strsplit_customv0}
\title{Split strings and retrieve a specific element}
\usage{
strsplit_customv0(infunc_list, infunc_split_by = "_", retreiving_onject_id = 1)
}
\arguments{
\item{infunc_list}{Character vector. Strings to split.}

\item{infunc_split_by}{Character. Delimiter. Default = "_".}

\item{retreiving_onject_id}{Integer. Position to retrieve after splitting. Default = 1.}
}
\value{
Character vector of retrieved elements.
}
\description{
Splits each string by a delimiter and returns the element at the specified position.
}
